python/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docs-2-cms"
version = "0.1.0"
description = "A simple utility for publishing git-managed markdown files to a WordPress site via API"
readme = "../README.md"
requires-python = ">=3.14"

dependencies = [
    "httpx>=0.27.0",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.0",
    "python-dotenv>=1.0.1",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10",
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5",
    "respx>=0.21.1",
]

[project.urls]
repository = "https://github.com/Dyc3r/docs2cms"
issues = "https://github.com/Dyc3r/docs2cms/issues"

[project.scripts]
d2cms = "d2cms.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
pythonpath = ["src"]

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true

pretty = true
show_error_codes = true
